requirements.txt
Django==4.1.5
django-bootstrap4==22.3
python-dotenv==0.21.1
requests == 2.28.2
boto3 == 1.26.69
